Move translation from rule to match into the BaseRule trait
