+++ app/matchers/RegexMatcher.scala
@@ -29,12 +29,10 @@ class RegexMatcher(category: String, rules: List[RegexRule]) extends Matcher {
   private def checkRule(request: MatcherRequest, rule: RegexRule): List[RuleMatch] = {
     request.blocks.flatMap { block =>
         rule.regex.findAllMatchIn(block.text).map { currentMatch =>
-          println(currentMatch)
-          RuleMatch.fromRegex(
+          rule.toMatch(
             currentMatch.start,
             currentMatch.end,
-            block,
-            rule
+            block
           ) }
     }
   }
