+++ app/model/RegexRule.scala
@@ -16,4 +16,21 @@ case class RegexRule(
     suggestions: List[TextSuggestion] = List.empty,
     replacement: Option[TextSuggestion] = None,
     regex: Regex
-) extends BaseRule
+) extends BaseRule {
+
+  def toMatch(start: Int, end: Int, block: TextBlock): RuleMatch = {
+    val matchedText = block.text.substring(start, end)
+    RuleMatch(
+      rule = this,
+      fromPos = start + block.from,
+      toPos = end + block.from,
+      matchedText = matchedText,
+      message = description,
+      shortMessage = Some(description),
+      suggestions = suggestions,
+      replacement = replacement.map(replacement => regex.replaceAllIn(matchedText, replacement.text)),
+      markAsCorrect = replacement.map(_.text).getOrElse("") == block.text.substring(start, end),
+      matchContext = surroundingText(block.text, start, end, surroundingBuffer)
+    )
+  }
+}
